+++ tests/Unit/Private/DSC/GetDscResourceDetail.tests.ps1
@@ -43,27 +43,6 @@ Describe 'GetDscResourceDetail' {
         }
     }
 
-    Context 'Custom Path' {
-        It 'Works with a custom path' {
-            InModuleScope -ScriptBlock {
-                if (TestWinGetModule)
-                {
-                    # TODO: life is difficult with WinGet
-                    $version = (GetDscVersion) -replace "preview.", ""
-                    $architecture = ([System.Runtime.InteropServices.RuntimeInformation]::OSArchitecture).ToString().ToLower()
-                    $Path = Join-Path $env:ProgramFiles 'WindowsApps' "Microsoft.DesiredStateConfiguration-Preview_3.0.$version.0_$architecture`__8wekyb3d8bbwe" 'dsc.exe'
-                    if ($env:TF_BUILD)
-                    {
-                        $Path = Join-path $env:ProgramFiles 'dsc' 'dsc.exe'
-                    }
-                }
-                $result = GetDscResourceDetail -Path $Path
-                $result | Should -Not -BeNullOrEmpty
-            }
-
-        }
-    }
-
     Context 'Invalid Path' {
         It 'Throws an error when the path does not exist' {
             InModuleScope -ScriptBlock {