+++ CHANGELOG.md
@@ -5,6 +5,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed 
+
+- Added Linux and UseGitHub option when installing `dsc.exe`
+
+## [1.0.1] - 2024-11-04
+
+### Added
+
+- Official release with WinGet
+
 ## [1.0.0] - 2024-10-04
 
 ### Added